+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2025-09-30
+
 ### Added
 
 - Support for prompt arguments with variable replacement using `{variable}` format in templates
 - Escape literal brackets using double brackets (`{{var}}`)
 - Support for `title` field in frontmatter (defaults to filename)
+
+### Changed
+
+- Remove tag `local` from all prompts, add tag `shinkuro`.
 
 ## [0.2.0] - 2025-09-29
 
@@ -28,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Local file mode
 - GitHub mode
 
-[unreleased]: https://github.com/DiscreteTom/shinkuro/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/DiscreteTom/shinkuro/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/DiscreteTom/shinkuro/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/DiscreteTom/shinkuro/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/DiscreteTom/shinkuro/releases/tag/v0.1.0